refactor fp. easier to implement new fp methods
