+++ dpgen2/fp/__init__.py
@@ -1,1 +1,13 @@
-from .vasp import VaspInputs
+from .vasp import (
+    VaspInputs,
+    PrepVasp,
+    RunVasp,
+)
+
+fp_styles = {
+    "vasp" :  {
+        "inputs" : VaspInputs,
+        "prep" : PrepVasp,
+        "run" : RunVasp,
+    }
+}
