+++ dpgen2/fp/vasp.py
@@ -1,5 +1,3 @@
-import numpy as np
-import dpdata
 from pathlib import Path
 from typing import (
     Optional,
@@ -9,113 +7,88 @@ from typing import (
     Dict,
     Union,
 )
+import numpy as np
+import dpdata
+from dargs import (
+    dargs, 
+    Argument, 
+    Variant, 
+    ArgumentEncoder,
+)
 
-class VaspInputs():
-    def __init__(
-            self,
-            kspacing : Union[float, List[float]],
-            incar_template_name : str,
-            potcar_names : Dict[str, str],
-            kgamma : bool = True,
+from .prep_fp import PrepFp
+from .run_fp import RunFp
+from .vasp_input import VaspInputs, make_kspacing_kpoints
+
+
+# global static variables
+vasp_conf_name = 'POSCAR'
+vasp_input_name = 'INCAR'
+vasp_pot_name = 'POTCAR'
+vasp_kp_name = 'KPOINTS'
+
+
+class PrepVasp(PrepFp):
+    def prep_task(
+            conf_frame: dpdata.System,
+            vasp_inputs: VaspInputs,
     ):
-        """
-        Parameters
-        ----------
-        kspacing : Union[float, List[float]]
-                The kspacing. If it is a number, then three directions use the same
-                ksapcing, otherwise it is a list of three numbers, specifying the
-                kspacing used in the x, y and z dimension.
-        incar_template_name: str
-                A template INCAR file. 
-        potcar_names : Dict[str,str]
-                The potcar files for the elements. For example
-                { 
-                   "H" : "/path/to/POTCAR_H",
-                   "O" : "/path/to/POTCAR_O",
-                }
-        kgamma : bool
-                K-mesh includes the gamma point
-        """
-        self.kspacing = kspacing
-        self.kgamma = kgamma
-        self.incar_from_file(incar_template_name)
-        self.potcars_from_file(potcar_names)
+        conf_frame.to('vasp/poscar', vasp_conf_name)
+        Path(vasp_input_name).write_text(
+            vasp_inputs.incar_template
+        )
+        # fix the case when some element have 0 atom, e.g. H0O2
+        tmp_frame = dpdata.System(vasp_conf_name, fmt='vasp/poscar')
+        Path(vasp_pot_name).write_text(
+            vasp_inputs.make_potcar(tmp_frame['atom_names'])
+        )
+        Path(vasp_kp_name).write_text(
+            vasp_inputs.make_kpoints(conf_frame['cells'][0])
+        )
 
-    @property
-    def incar_template(self):
-        return self._incar_template
+        
+class RunVasp(RunFp):
+    def input_files() -> List[str]:
+        return [vasp_conf_name, vasp_input_name, vasp_pot_name, vasp_kp_name]
 
-    @property
-    def potcars(self):
-        return self._potcars
+    def optional_input_files() -> List[str]:
+        return []
 
-    def incar_from_file(
-            self,
-            fname : str,
-    ):
-        self._incar_template = Path(fname).read_text()
-
-    def potcars_from_file(
-            self,
-            dict_fnames : Dict[str,str],
-    ):
-        self._potcars = {}
-        for kk,vv in dict_fnames.items():
-            self._potcars[kk] = Path(vv).read_text()            
-
-    def make_potcar(
-            self, 
-            atom_names,
-    ) -> str:        
-        potcar_contents = []
-        for nn in atom_names:
-            potcar_contents.append(self._potcars[nn])
-        return "".join(potcar_contents)            
-
-    def make_kpoints(
-            self,
-            box : np.ndarray,
-    ) -> str:
-        return make_kspacing_kpoints(box, self.kspacing, self.kgamma)
+    def run_task(
+            command : str,
+            log_name: str,
+            out_name: str,
+    ) -> Tuple[str, str]:
+        # run vasp
+        command = ' '.join([command, '>', log_name])
+        ret, out, err = run_command(command, shell=True)
+        if ret != 0:
+            raise TransientError(
+                'vasp failed\n',
+                'out msg', out, '\n',
+                'err msg', err, '\n'
+            )                    
+        # convert the output to deepmd/npy format
+        sys = dpdata.LabeledSystem('OUTCAR')
+        sys.to('deepmd/npy', out_name)
+        return out_name, log_name
 
 
-def make_kspacing_kpoints(box, kspacing, kgamma) :
-    if type(kspacing) is not list:
-        kspacing = [kspacing, kspacing, kspacing]
-    box = np.array(box)
-    rbox = _reciprocal_box(box)
-    kpoints = [max(1,(np.ceil(2 * np.pi * np.linalg.norm(ii) / ks).astype(int))) for ii,ks in zip(rbox,kspacing)]
-    ret = _make_vasp_kpoints(kpoints, kgamma)
-    return ret
+    @staticmethod
+    def args():
+        doc_vasp_cmd = "The command of VASP"
+        doc_vasp_log = "The log file name of VASP"
+        doc_vasp_out = "The output dir name of labeled data. In `deepmd/npy` format provided by `dpdata`."
+        return [
+            Argument("command", str, optional=True, default='vasp', doc=doc_vasp_cmd),
+            Argument("log", str, optional=True, default=vasp_default_log_name, doc=doc_vasp_log),
+            Argument("out", str, optional=True, default=vasp_default_out_data_name, doc=doc_vasp_out),
+        ]
 
-
-def _make_vasp_kp_gamma(kpoints):
-    ret = ""
-    ret += "Automatic mesh\n"
-    ret += "0\n"
-    ret += "Gamma\n"
-    ret += "%d %d %d\n" % (kpoints[0], kpoints[1], kpoints[2])
-    ret += "0  0  0\n"
-    return ret
-
-def _make_vasp_kp_mp(kpoints):
-    ret = ""
-    ret += "K-Points\n"
-    ret += "0\n"
-    ret += "Monkhorst Pack\n"
-    ret += "%d %d %d\n" % (kpoints[0], kpoints[1], kpoints[2])
-    ret += "0  0  0\n"
-    return ret
-
-def _make_vasp_kpoints (kpoints, kgamma = False) :
-    if kgamma :
-        ret = _make_vasp_kp_gamma(kpoints)
-    else :
-        ret = _make_vasp_kp_mp(kpoints)
-    return ret
-    
-def _reciprocal_box(box) :
-    rbox = np.linalg.inv(box)
-    rbox = rbox.T
-    return rbox
-
+    @staticmethod
+    def normalize_config(data = {}, strict=True):
+        ta = RunVasp.vasp_args()
+        base = Argument("base", dict, ta)
+        data = base.normalize_value(data, trim_pattern="_*")
+        base.check_value(data, strict=strict)
+        return data
