+++ dpgen2/op/__init__.py
@@ -2,7 +2,5 @@ from .prep_dp_train import PrepDPTrain
 from .run_dp_train import RunDPTrain
 from .prep_lmp import PrepLmp
 from .run_lmp import RunLmp
-from .prep_vasp import PrepVasp
-from .run_vasp import RunVasp
 from .collect_data import CollectData
 from .select_confs import SelectConfs
